# enhance: Optimize save colelction target latency