issue: #38237
this PR only use better compression level for proto msg which is larger than 1MB, and use a lighter compression level for smaller proto msg, which could get a better latency in most case.

this PR could reduce the latency from 22.7s to 4.7s with 10000 collctions and each collections has 1000 segments.

before this PR:
BenchmarkTargetManager-8               1        22781536357 ns/op       566407275088 B/op       11188282 allocs/op
after this PR:
BenchmarkTargetManager-8               1        4729566944 ns/op        36713248864 B/op        10963615 allocs/op